Move v0.4 contracts and tests over
