+++ evm-contracts/src/v0.4/Oracle.sol
