+++ evm-contracts/src/v0.4/Pointer.sol
