+++ evm-contracts/src/v0.4/interfaces/AggregatorInterface.sol
