+++ evm-contracts/src/v0.4/interfaces/OracleInterface.sol
