+++ evm-contracts/src/v0.4/tests/EmptyOracle.sol
