+++ evm-contracts/src/v0.4/tests/MaliciousConsumer.sol
