+++ evm-contracts/test/v0.4/BasicConsumer.test.ts
@@ -8,8 +8,8 @@ import {
 import cbor from 'cbor'
 import { assert } from 'chai'
 import { ethers } from 'ethers'
-import { BasicConsumerFactory } from '../src/generated/BasicConsumerFactory'
-import { OracleFactory } from '../src/generated/OracleFactory'
+import { BasicConsumerFactory } from '../../ethers/v0.4/BasicConsumerFactory'
+import { OracleFactory } from '../../ethers/v0.4/OracleFactory'
 
 const basicConsumerFactory = new BasicConsumerFactory()
 const oracleFactory = new OracleFactory()
