+++ evm-contracts/test/v0.4/Pointer.test.ts
@@ -1,6 +1,6 @@
 import { contract, matchers, setup } from '@chainlink/test-helpers'
 import { assert } from 'chai'
-import { PointerFactory } from '../src/generated/PointerFactory'
+import { PointerFactory } from '../../ethers/v0.4/PointerFactory'
 
 const pointerFactory = new PointerFactory()
 const linkTokenFactory = new contract.LinkTokenFactory()
