+++ evm-test-helpers/package.json
@@ -24,7 +24,7 @@
     "chai": "^4.2.0",
     "chalk": "^2.4.2",
     "debug": "^4.1.1",
-    "ethers": "^4.0.41"
+    "ethers": "^4.0.44"
   },
   "devDependencies": {
     "@types/cbor": "^5.0.0",
